ID review edits
We try to avoid referring to the guide itself, so could we point out the location of the directory in a different way? "The 'finish' directory in the root of this guide contains..."?

I'm wondering if we should display some of the info for @Timed, @Counted, @Guage, as well as the test cases, in a table.